+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/ShootPatterns/Scripts/BasicShotPattern.cs
@@ -5,20 +5,18 @@ using System.Threading;
 [CreateAssetMenu(fileName = "BASIC_", menuName = "Touhou Spell Up/Danmaku/Bullet Pattern/Basic Shot")]
 public class BasicShotPattern : ShootPatternBase
 {
-    public override async UniTask ExecuteImpl(Transform spawnPoint, Bullet bulletToUse, CancellationToken token)
+    public override async UniTask ExecuteImpl(Shooter shooter, Bullet bulletToUse, CancellationToken token)
     {
         if (bulletToUse == null || bulletToUse.Prefab == null)
         {
             Debug.LogError("発射する弾が指定されていません！", this);
             return;
         }
+        Vector3 spawnPosition = GetSpawnPosition(shooter);
+        Quaternion spawnRotation = shooter.transform.rotation;
+
         // spawnPointの位置と角度で、指定された弾を1つ生成する
-        var bulletInstance = Instantiate(bulletToUse.Prefab, spawnPoint.position, spawnPoint.rotation);
-        var enemyBullet = bulletInstance.GetComponent<EnemyBullet>();
-        if (enemyBullet != null)
-        {
-            enemyBullet.Initialize(bulletToUse.Property);
-        }
+        shooter.InstantiateBullet(bulletToUse, spawnPosition, spawnRotation);
 
         await UniTask.CompletedTask;
     }